+++ libs/nest/prisma/migrations/20240117173313_add_compliance_survey_type/migration.sql
@@ -1,2 +1,0 @@
--- AlterEnum
-ALTER TYPE "survey_types" ADD VALUE 'COMPLIANCE';